src/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
numpy==1.21.0
Pillow==8.3.1
pyfiglet==0.8.post1
requests==2.26.0
urllib3==1.26.6
